requirements.txt
crewai==1.1.0
crewai-tools==1.1.0
langchain==0.1.20
langchain-openai==0.1.7
langchain-ollama==0.1.3
langchain-community==0.0.38
python-dotenv>=1.1.1
faiss-cpu==1.12.0
litellm==1.39.2
anthropic==0.27.0
fastapi==0.110.2
uvicorn[standard]==0.29.0
jinja2==3.1.3
pytest==8.1.1
scikit-learn>=1.3.0
